Разделение на модули v2
